make into util function
